Explicit cast to (size_t)
P->wlen is set to be PCRYPT_ESZ (:= 1060) or the returned value of
proto_crypt_dec().  That function promises to
    "decrypt PCRYPT_ESZ bytes... and return the length; otherwise, return -1"
although upon reading the source of proto_crypt_dec(), we see that if it does
not return -1, the value will be
    0 < len <= PCRYPT_MAXDSZ
As such, the cast is safe.

+++ proto/proto_pipe.c
@@ -109,11 +109,13 @@ callback_pipe_read(void * cookie, ssize_t len)
 	/* Write the encrypted or decrypted data. */
 	if (P->decr) {
 		if ((P->write_cookie = network_write(P->s_out, P->dbuf,
-		    P->wlen, P->wlen, callback_pipe_write, P)) == NULL)
+		    (size_t)P->wlen, (size_t)P->wlen, callback_pipe_write,
+		    P)) == NULL)
 			goto err0;
 	} else {
 		if ((P->write_cookie = network_write(P->s_out, P->ebuf,
-		    P->wlen, P->wlen, callback_pipe_write, P)) == NULL)
+		    (size_t)P->wlen, (size_t)P->wlen, callback_pipe_write,
+		    P)) == NULL)
 			goto err0;
 	}
